move utility functions to utils module
